Fixed pipes.
Fixed pipe read and EOF bugs.
Fixed printInt with negative numbers.

+++ SharedLib/myUtils.h
@@ -5,5 +5,5 @@
 
 void numToString(int num, uint8_t digits, char out[], uint8_t base);
 uint64_t pow(unsigned base, unsigned exponent);
-uint8_t countDigits(unsigned num, uint8_t base);
+unsigned countDigits(int num, unsigned base);
 void unsignedToString(uint64_t num, uint8_t digits, char out[], uint8_t base);
